make the library use float tensor type to avoid exceptions. Not sure why this is different from optimise_* or main.py scripts though

+++ allagents/agents/cacla.py
@@ -52,8 +52,8 @@ class CACLA(Agent):
         self._features = None
         self._action = None
         self._state = None
-        self._action_low = torch.tensor(action_space.low, device=policy.device)
-        self._action_high = torch.tensor(action_space.high, device=policy.device)
+        self._action_low = torch.tensor(action_space.low, device=policy.device).float()
+        self._action_high = torch.tensor(action_space.high, device=policy.device).float()
 
     def _normal(self, output):
         if self._log:
@@ -76,7 +76,7 @@ class CACLA(Agent):
         # self.writer.add_scalar("action/det", deterministic_action)
 
         # Get the stochastic action by centering a Normal distribution on the policy output
-        stochastic_action = self._normal(deterministic_action).sample()
+        stochastic_action = self._normal(deterministic_action).sample().float()
 
         # Clip the stochastic action to the gym environment's action space
         stochastic_action = torch.max(stochastic_action, self._action_low)
